staking primitives
