+++ primitives/src/dapp_staking.rs
@@ -1,0 +1,81 @@
+// This file is part of Astar.
+
+// Copyright (C) 2019-2023 Stake Technologies Pte.Ltd.
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+// Astar is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// Astar is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with Astar. If not, see <http://www.gnu.org/licenses/>.
+
+use super::Balance;
+
+// TODO2: However this ends up looking in the end, we should not duplicate these parameters in the runtime.
+//        Both the dApp staking & inflation pallet should use the same source.
+/// TODO: docs!
+pub trait CycleConfiguration {
+    /// How many different periods are there in a cycle (a 'year').
+    ///
+    /// This value has to be at least 1.
+    fn periods_per_cycle() -> u32;
+
+    /// For how many standard era lengths does the voting subperiod last.
+    ///
+    /// This value has to be at least 1.
+    fn eras_per_voting_subperiod() -> u32;
+
+    /// How many standard eras are there in the build&earn subperiod.
+    ///
+    /// This value has to be at least 1.
+    fn eras_per_build_and_earn_subperiod() -> u32;
+
+    /// How many blocks are there per standard era.
+    ///
+    /// This value has to be at least 1.
+    fn blocks_per_era() -> u32;
+
+    /// For how many standard era lengths does the period last.
+    fn eras_per_period() -> u32 {
+        Self::eras_per_voting_subperiod().saturating_add(Self::eras_per_build_and_earn_subperiod())
+    }
+
+    /// For how many standard era lengths does the cylce (a 'year') last.
+    fn eras_per_cycle() -> u32 {
+        Self::eras_per_period().saturating_mul(Self::periods_per_cycle())
+    }
+
+    /// How many blocks are there per cycle (a 'year').
+    fn blocks_per_cycle() -> u32 {
+        Self::blocks_per_era().saturating_mul(Self::eras_per_cycle())
+    }
+
+    /// For how many standard era lengths do all the build&earn subperiods in a cycle last.    
+    fn build_and_earn_eras_per_cycle() -> u32 {
+        Self::eras_per_build_and_earn_subperiod().saturating_mul(Self::periods_per_cycle())
+    }
+}
+
+/// Interface for staking reward handler.
+///
+/// Provides reward pool values for stakers - normal & bonus rewards, as well as dApp reward pool.
+/// Also provides a safe function for paying out rewards.
+pub trait StakingRewardHandler<AccountId> {
+    /// Returns the staker reward pool & dApp reward pool for an era.
+    ///
+    /// The total staker reward pool is dynamic and depends on the total value staked.
+    fn staker_and_dapp_reward_pools(total_value_staked: Balance) -> (Balance, Balance);
+
+    /// Returns the bonus reward pool for a period.
+    fn bonus_reward_pool() -> Balance;
+
+    /// Attempts to pay out the rewards to the beneficiary.
+    fn payout_reward(beneficiary: &AccountId, reward: Balance) -> Result<(), ()>;
+}
