+++ primitives/src/lib.rs
@@ -35,6 +35,9 @@ pub mod evm;
 /// Precompiles
 pub mod precompiles;
 
+/// dApp staking & inflation primitives.
+pub mod dapp_staking;
+
 /// Useful primitives for testing.
 pub mod testing;
